Change infrastructure to allow to devide context in multiple files
