--- KRFHomepage.Domain/Database/Translation/Token.cs
@@ -1,12 +1,0 @@
-﻿using System.Collections.Generic;
-using System.ComponentModel.DataAnnotations;
-
-namespace KRFHomepage.Domain.Database.Translation
-{
-    public class Token
-    {
-        [Key]
-        public string Value { get; set; }
-        public virtual ICollection<Translation> Translations { get; set; }
-    }
-}
