+++ KRFHomepage.Infrastructure/Database/DBContext/HomepageDBContext.cs
@@ -1,0 +1,20 @@
+﻿using Microsoft.EntityFrameworkCore;
+
+namespace KRFHomepage.Infrastructure.Database.DBContext
+{
+    public partial class HomepageDBContext : DbContext
+    {
+        public HomepageDBContext(DbContextOptions options) : base(options)
+        {
+        }
+        protected override void OnConfiguring(DbContextOptionsBuilder optionsBuilder)
+        {
+        }
+
+        protected override void OnModelCreating(ModelBuilder modelBuilder)
+        {
+            this.CommonModelCreating(modelBuilder);
+            this.TranslationModelCreating(modelBuilder);
+        }
+    }
+}
